requirements.txt
botbuilder-integration-aiohttp>=4.12.0
requests